update - being, weapon render

+++ Projectile.h
@@ -3,9 +3,9 @@
 
 void DestroyProjectiles(Projectiles_array* const);
 static bool ProjectileHitsBeing(Projectile* const, Being* const);
-void AddPCProjectileToArray(Projectiles_array* const, const SDL_FPoint* const, const float, const float, const Impact* const, const unsigned int);
+void AddPCProjectileToArray(Projectiles_array* const, const SDL_FPoint* const, const float, const float, const Impact* const, const Uint8);
 void AddHProjectileToArray(Projectiles_array* const, const SDL_FPoint* const, const float, const float, const Impact* const);
-static void DestroyProjectileInArray(Projectiles_array* const, const unsigned int);
+static void DestroyProjectileInArray(Projectiles_array* const, const Uint16);
 static void MoveProjectile(Projectile* const);
 void UpdateProjectiles(Game_data* const);
 static bool ProjectileHitsPlayer(Projectile* const, Player* const);
